Commiting changes
