+++ tests/test_git_parser.py
@@ -34,3 +34,46 @@ def test_get_commit_contents(repository_paths: dict[str, pathlib.Path]) -> None:
         commit_contents = get_commit_contents(str(repo_path), commit.hexsha)
         # Ensure the result is a dictionary
         assert isinstance(commit_contents, dict)
+
+
+"""
+Testing LoC_tracker functionality for calculating total lines changed in a given repository.
+"""
+
+import pathlib
+
+
+from almanack.git_parser import get_commit_logs, calculate_loc_changes
+
+
+def get_most_recent_commits(repo_path: pathlib.Path) -> tuple[str, str]:
+    """
+    Determine the two most recent commit hashes in the test repositories.
+
+    Args:
+        repo_path (pathlib.Path): The path to the git repository.
+
+    Returns:
+        tuple[str, str]: Tuple containing the source and target commit hashes.
+    """
+    commit_logs = get_commit_logs(repo_path)
+
+    # Sort commits by their timestamp to get the two most recent ones
+    sorted_commits = sorted(commit_logs.items(), key=lambda item: item[1]["timestamp"])
+
+    # Get the commit hashes of the two most recent commits
+    source_commit = sorted_commits[-2][0]
+    target_commit = sorted_commits[-1][0]
+
+    return source_commit, target_commit
+
+
+def test_calculate_loc_changes(repository_paths: dict[str, pathlib.Path]) -> None:
+    """
+    Test the calculate_loc_changes function
+    """
+    for _, repo_path in repository_paths.items():
+        source_commit, target_commit = get_most_recent_commits(repo_path)
+        # Calculate lines of code changes between source and target commits
+        print(calculate_loc_changes(repo_path, source_commit, target_commit))
+
